feat(validate-links-sts): Refactor and Document
Learning, that we cannot fully optimize loading into just any Excel
(indepentent of the OS), skipped the parts that makes it harder for
other tools to process the CSV. Instead, focus in the validity across
consuming applications and instead document, how to it is recommended to
be loaded to Excel.

Also, introduces a new suggestion for a structure of the documentation
with each tool coming with its own documentation.